+++ backend/models/review.py
@@ -4,9 +4,9 @@ from datetime import datetime
 class Review(db.Model):
     __tablename__ = "reviews"
 
-    id = db.Column(db.BigInteger, primary_key=True, autoincrement=True)
-    application_id = db.Column(db.BigInteger, db.ForeignKey("applications.id"), nullable=False)
-    supervisor_id = db.Column(db.BigInteger, db.ForeignKey("users.id"), nullable=False)
+    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
+    application_id = db.Column(db.Integer, db.ForeignKey("applications.id"), nullable=False)
+    supervisor_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     recommendation = db.Column(db.String(50), nullable=True)
     remarks = db.Column(db.Text, nullable=True)
     reviewed_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=True)
